+++ spec/controllers/api/event_hours_controller_spec.rb
@@ -112,4 +112,14 @@ describe Api::EventHoursController, type: :controller do
       expect(JSON.parse(response.body).size).to eq(1)
     end
   end
+
+  def format_time_key(time_key)
+    time_key = "0#{time_key}" if time_key.length == 3
+    minutes = time_key.last(2)
+    if minutes == '00'
+      Time.strptime(time_key, '%H%M').strftime('%l %p').lstrip
+    else
+      Time.strptime(time_key, '%H%M').strftime('%l:%M %p').lstrip
+    end
+  end
 end
